requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.124.4
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.237
ultralytics-thop==2.0.18
urllib3==2.6.2
uvicorn==0.38.0
